% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_information.R
\name{aggregation_information}
\alias{aggregation_information}
\title{aggregation_information}
\usage{
aggregation_information(timeshift, aggregation_interval)
}
\arguments{
\item{timeshift}{timeshift which is applied to the data}

\item{aggregation_interval}{aggregation interval for the modelling, for more
information about the aggregation_interval units see
?lubridate::round_date}
}
\value{
an object of the class `stAirPol.aggregation_information`
}
\description{
aggregation_information
}
\examples{
m.agg_infor <- aggregation_information(timeshift = lubridate::hours(2),
aggregation_interval = '8 hours')
print(m.agg_infor)
}
