% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_hist.R
\name{calculate_hist_grid_points}
\alias{calculate_hist_grid_points}
\title{calculate_hist_grid_points}
\usage{
calculate_hist_grid_points(data, lambda, lag)
}
\arguments{
\item{data}{a object as returned by \link{smooth_space_time_variables}, but
splitted up by corrdinated}

\item{lambda}{double smoothing parameter lambda}

\item{lag}{integer lag parameter}
}
\value{
data.frame with the additional column hist
}
\description{
calculate for each point the transformated wind and rain data
}
