% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_space_time_datasets.R
\name{calculate_space_time_datasets}
\alias{calculate_space_time_datasets}
\title{Calculate datasets and grids for humidity and temperature}
\usage{
calculate_space_time_datasets()
}
\value{
nothing, but 8 datasets get assigned to the global enviroment
with the names: data.humi.p1, data.humi.p2, grid.humi.p1, grid.humi.p2,
data.temp.p1, data.temp.p2, grid.temp.p1 and grid.temp.p2
}
\description{
Calculate datasets and grids for humidity and temperature
}
