% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_time_datasets.R
\name{calculate_time_datasets}
\alias{calculate_time_datasets}
\title{Calculate datasets and grids for wind and rain}
\usage{
calculate_time_datasets()
}
\value{
nothing, but 16 datasets get assigned to the global enviroment
with the names:
data.wind.p1, data.wind.p2, grid.wind.p1, grid.wind.p2,
data.rain.p1, data.rain.p2, grid.rain.p1, grid.rain.p2,
data.windhist.p1, data.windhist.p2, grid.windhist.p1, grid.windhist.p2,
data.rainhist.p1, data.rainhist.p2, grid.rainhist.p1 and grid.rainhist.p2
}
\description{
Calculate datasets and grids for wind and rain
}
