% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_space_time_datasets.R
\name{check_hist}
\alias{check_hist}
\title{Check the output of smooth_space_time_variables}
\usage{
check_hist(data)
}
\arguments{
\item{data}{\link{data.table} which contains a column with the name prediction}
}
\value{
a warnings if any NAs are detected in the column prediction
}
\description{
Check the output of smooth_space_time_variables
}
\examples{
df <-  data.frame(prediction =c (1, 1, NA))
check_prediction(df)
}
