% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_data.R
\name{clean_model_data}
\alias{clean_model_data}
\title{clean_model_data}
\usage{
clean_model_data(model_data, timesIQR = 1.5)
}
\arguments{
\item{model_data}{Dataset as a data.table which one time dimension in the
variable timestamp and one numeric varibale with the name values}

\item{times_IQR}{remove outliers,which are more then times_IQR * IQR away
from the median}
}
\value{
model_data, with all values, which are outside the timesIQR plus minus
the median, are in one timestamp set to NA
}
\description{
Remove outliers from the dataset. If an outlier is detected, the entry in
the column value is set to NA.
}
