% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLite.R
\name{connection}
\alias{connection}
\title{SQLite connection}
\usage{
connection(date_pattern, variable_type, path)
}
\arguments{
\item{date_pattern}{date pattern which specifies the month in the format 'YYYY_MM'}

\item{variable_type}{which measured variable , P1, P2, pressure, temperature, humidity}

\item{path}{filepath where the informations are stored}
}
\value{
a RSQLite DB connection
}
\description{
Builds an SQLite connection to the specified database.
}
\examples{
connection('2018_01', 'P1')
}
