% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalulate_prediction.R
\name{Coverage}
\alias{Coverage}
\title{Coverage}
\usage{
Coverage(zhat, up, low)
}
\arguments{
\item{zhat}{predicted values}

\item{up}{upper quantile}

\item{low}{lower quantile}
}
\description{
Coverage
}
\examples{
data("mini_dataset")
mini_dataset <- clean_model_data(mini_dataset)
formula = value ~ humi + temp + rainhist + windhist +
  trafficvol + log(sensor_age)
training_set <- get_test_and_training_set(mini_dataset, sampel_size = 0.75,
                                          random.seed = 220292)
model.gp <- fit_sp_model(data = mini_dataset, formula = formula,
                         model = 'GP', training_set = training_set)
pred.gp <- predict(model.gp, mini_dataset, training_set)
Coverage(pred.gp$value, pred.gp$up, pred.gp$low)
}
