% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dvd_data.R
\name{download_dwd_data}
\alias{download_dwd_data}
\title{Download DWD Data}
\usage{
download_dwd_data(obs, path, type = "recent")
}
\arguments{
\item{obs}{which observation should be downloaded? See
\url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/hourly/} for
more informations}

\item{path}{filepath where the informations should be stored}

\item{type}{type = 'recent' for about the last two years, type = 'historical'
for a longer historical timeframe}
}
\description{
That functions downloads all recent informations from the DWD
(Deutscher Wetter Dienst)
}
\examples{
download_dvd_data(obs = 'wind')
download_dvd_data(obs = 'precipitation')
}
