% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_open_traffic_map.R
\name{download_OTM}
\alias{download_OTM}
\title{Download Open Transport Map Data}
\usage{
download_OTM(path)
}
\arguments{
\item{path}{filepath where the informations should be stored}
}
\description{
That functions downloads all informations about traffic volumen in Germany
}
\examples{
download_OTM(path = './')
}
