% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sensor_age.R
\name{download_sensor_age}
\alias{download_sensor_age}
\title{Setup informations about the age of each sensor}
\usage{
download_sensor_age(path)
}
\arguments{
\item{path}{filepath where the informations should be stored}
}
\description{
The functions stores a csv file in the selected path
}
\examples{
sensor_age(path = './')
}
