% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid_traffic.R
\name{estimate_grid_size}
\alias{estimate_grid_size}
\title{estimate_grid_size}
\usage{
estimate_grid_size(plz, grid_cellsize)
}
\arguments{
\item{plz}{a character vector which contains all german postcodes of the
area which should be used.}

\item{grid_cellsize}{\link{sf::st_make_grid} for informations about the
cellsize parameter}
}
\description{
Estimate the size of the choosen grid parameters
}
\examples{
estimate_grid_size('80993', 0.01)
}
