% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalulate_prediction.R
\name{evaluate_prediction}
\alias{evaluate_prediction}
\title{evaluate_prediction}
\usage{
evaluate_prediction(pred, round = 2, ...)
}
\arguments{
\item{pred}{an object as returned by \link{predict.stAirPol.model}}

\item{round}{digits for round the measures}

\item{...}{}
}
\description{
evaluate_prediction
}
\examples{
data("mini_dataset")
mini_dataset <- clean_model_data(mini_dataset)
formula = value ~ humi + temp + rainhist + windhist +
  trafficvol + log(sensor_age)
training_set <- get_test_and_training_set(mini_dataset, sampel_size = 0.75,
                                          random.seed = 220292)
model.gp <- fit_sp_model(data = mini_dataset, formula = formula,
                         model = 'GP', training_set = training_set)
pred.gp <- predict(model.gp, mini_dataset, training_set)
evaluate_prediction(pred.gp)
}
