% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_prediction_table.R
\name{evaluate_prediction_table}
\alias{evaluate_prediction_table}
\title{evaluate_prediction_table}
\usage{
evaluate_prediction_table(list)
}
\arguments{
\item{list}{a named list with elements of objects as returned by
\link{predict.stAirPol.model}}
}
\value{
a data.table
}
\description{
Formats a table of different prediction data.frames
}
