% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_subintervalls.R
\name{fit_subintervalls}
\alias{fit_subintervalls}
\title{Title}
\usage{
fit_subintervalls(data, grid = NULL, training_set = NULL, unit = "1 week",
  nItr = 5000, nBurn = 1000, return = "prediction", tol = 0.01,
  retry_count = 5, mc.cores = parallel::detectCores() - 1, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
retry model fitting and prediction in cause of a strange behavior
sometimes
}
