% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dt_to_points.R
\name{format_dt_to_points}
\alias{format_dt_to_points}
\title{format_dt_to_points}
\usage{
format_dt_to_points(dt)
}
\arguments{
\item{dt}{a datatable contains lon and lat as variable}
}
\value{
Geometry set of all points
}
\description{
Format a data.table with the columns lon and lat to a set of \link{sf::st_point}
}
