% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_frame.R
\name{get_model_frame}
\alias{get_model_frame}
\title{Combine datasets to a model frame}
\usage{
get_model_frame(sensor_data, sensors, expand = FALSE)
}
\arguments{
\item{sensor_data}{data as returned by \link{get_sensor_measured_values}}

\item{sensors}{data as returned by \link{get_sensors}}

\item{expand}{should the spatio and temporal dimensions expand that all
missing observations are filled by NA_numeric}
}
\value{
a \link{data.table}
}
\description{
Combine datasets to a model frame
}
