% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nuts_for_plz.R
\name{get_nuts_for_plz}
\alias{get_nuts_for_plz}
\title{get_nuts_for_plz}
\usage{
get_nuts_for_plz(plz)
}
\arguments{
\item{plz}{numeric vector of German postcodes}
}
\value{
character vectors of unique NUTS_3
}
\description{
Map plz to NUTS_3
}
\examples{
get_nuts_for_plz(80993)
get_nuts_for_plz(c(80993, 83487))
}
