% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shape_for_plz.R
\name{get_plz_shape}
\alias{get_plz_shape}
\title{get_shape_for_plz}
\usage{
get_plz_shape(plz)
}
\arguments{
\item{plz}{a character vector which contains all German postcodes of the
area.}
}
\value{
shapefile for that German postcodes
}
\description{
get_shape_for_plz
}
\examples{
get_shape_for_plz(80993)
get_shape_for_plz(83487)
}
