% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensor_age.R
\name{get_sensor_age}
\alias{get_sensor_age}
\title{get_sensor_age}
\usage{
get_sensor_age(path, units = "weeks")
}
\arguments{
\item{path}{filepath where the informations are stored}

\item{units}{in which units should the age calculated}
}
\value{
a data.table which contains the age of a sensor in months and the
sensor_id
}
\description{
Informations about the age of a sensor in weeks
}
\examples{
get_sensor_age('./')
}
