% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensor_values.R
\name{get_sensor_measured_values}
\alias{get_sensor_measured_values}
\title{get_sensor_measured_values}
\usage{
get_sensor_measured_values(sensors, date_pattern, path)
}
\arguments{
\item{sensors}{object as returned from the function \link{get_sensors}}

\item{date_pattern}{character vector in the format YYYY_MM}

\item{path}{filepath where the informations are stored}
}
\value{
all measured values
}
\description{
get_sensor_measured_values
}
