% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensors.R
\name{get_sensors}
\alias{get_sensors}
\title{get_sensors}
\usage{
get_sensors(date_pattern, plz, path)
}
\arguments{
\item{date_pattern}{character vector in the format Year-Month}

\item{plz}{a character vector which contains all German postcodes of the
area which should be used.}

\item{path}{filepath where the informations are stored}
}
\value{
\link{data.table} with all sensors in that time period
}
\description{
Extract all information’s about the sensors which are located in the chosen
plz.
}
\examples{
get_sensors('2018-01', 80539)
}
