% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_st_predictions.R
\name{get_test_and_training_set}
\alias{get_test_and_training_set}
\title{get_test_and_training_set}
\usage{
get_test_and_training_set(sensors, sampel_size = 0.75, random.seed = NULL)
}
\arguments{
\item{sensors}{modeling \link{data.table}, should contain informations about
the sensor_id and the coordinated in the columns \code{lon} and \code{lat}}

\item{sampel_size}{double between 0 and 1, which specifies the percentage of
training sensors}

\item{random.seed}{for reproducibly reasons}
}
\description{
Split the sensors into a training a test set
}
\examples{
data("mini_dataset")
get_test_and_training_set(mini_dataset)
}
