% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dwd_data.R
\name{get_wind_data}
\alias{get_wind_data}
\title{get_wind_data}
\usage{
get_wind_data(date_pattern, dwd_stat, path)
}
\arguments{
\item{date_pattern}{date pattern which specifies the month in the format 'YYYY_MM'}

\item{dwd_stat}{Stations of the DWD as returned in the format of get_dwd_sensor_ids()}

\item{path}{filepath where the informations are stored}
}
\value{
historical wind data
}
\description{
read and transform the downloaded DWD data
}
