% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_luftdaten.R
\name{initializeDB}
\alias{initializeDB}
\title{Initialize Database}
\usage{
initializeDB(name, start, verbose = FALSE)
}
\arguments{
\item{name}{name of database}

\item{start}{start date, e.g. "2018-03-01"}

\item{verbose}{print additional information}
}
\description{
Initialize Database
}
