% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_luftdaten.R
\name{LD_download}
\alias{LD_download}
\title{Download csv files from archive.liftdaten.info}
\usage{
LD_download(folder, start = "2015-10-01", end = Sys.Date(),
  sensortype = "")
}
\arguments{
\item{folder}{folder to store csv}

\item{start}{start date, e.g. "2015-10-01"}

\item{end}{end data, e.g. Sys.Date()}
}
\value{
files are stored in folder/
}
\description{
Download csv files from archive.liftdaten.info
}
\author{
Volker Schmid
}
