% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_luftdaten.R
\name{LD_sql}
\alias{LD_sql}
\title{Write luftdaten to mysqlite database}
\usage{
LD_sql(db, end = Sys.Date() - 1,
  source = "/media/schmid/local/data/luftdaten/", save = TRUE,
  folder = "./", dbname = "db/luftdaten", verbose = FALSE)
}
\arguments{
\item{db}{mysqlite database}

\item{end}{enddate, e.g. "2019-12-31"}

\item{source}{folder with csv files}

\item{save}{boolean, should database be backuped monthly, 2: save daily}

\item{folder}{folder with database for backups}

\item{dbname}{character, file name of database (for backup)}

\item{verbose}{print more information}
}
\value{
no return
}
\description{
Write luftdaten to mysqlite database
}
\author{
Volker Schmid
}
