% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_luftdaten.R
\name{LD_sql2}
\alias{LD_sql2}
\title{Write luftdaten to mysqlite database}
\usage{
LD_sql2(start, end = Sys.Date() - 1,
  source = "/media/schmid/local/data/luftdaten/", sensortype = c(),
  dbname = "db/luftdaten", verbose = FALSE)
}
\arguments{
\item{start}{start date, e.g. "2016-10-01"}

\item{end}{enddate, e.g. "2019-12-31"}

\item{source}{folder with csv files}

\item{sensortype}{vector with sensortypes to process}

\item{dbname}{character, file name of database (for backup)}

\item{verbose}{print more information}
}
\value{
no return
}
\description{
Write luftdaten to mysqlite database
}
\author{
Volker Schmid
}
