% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_opentransportmap_data.R
\name{load_opentransportmap_data_nut3}
\alias{load_opentransportmap_data_nut3}
\title{load_opentransportmap_data_nut3}
\usage{
load_opentransportmap_data_nut3(nut3, path)
}
\arguments{
\item{nut3}{a character vector of nuts3, for more informations see
\url{https://de.wikipedia.org/wiki/NUTS}}
}
\value{
a sfc object of the roads
}
\description{
load_opentransportmap_data_nut3
}
\examples{
load_opentransportmap_data_nut3("DE212", "./")
}
