% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_traffic.R
\name{make_data_traffic}
\alias{make_data_traffic}
\title{make_data_traffic}
\usage{
make_data_traffic(sensors, lambda, scale = 100,
  mc.cores = parallel::detectCores() - 1)
}
\arguments{
\item{sensors}{data as returned by \link{get_sensors}}

\item{lambda}{smoothing parameter}

\item{scale}{numericvalue to scale the traffic volumen, defauld is 100}

\item{mc.cores}{how much cores should be used for parallelisation, default is
one core less you maximum number of detected cores.}
}
\value{
\link{data.table} of the sensors with a new column which contains
informations about the traffic volumen
}
\description{
Caclulate for each sensor the traffic volumen data
}
