% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid_traffic.R
\name{make_grid_traffic}
\alias{make_grid_traffic}
\title{make_grid}
\usage{
make_grid_traffic(lambda, plz, scale = 100,
  mc.cores = parallel::detectCores() - 1)
}
\arguments{
\item{lambda}{numerical smoothing parameter}

\item{plz}{numerical vector of german post codes}

\item{scale}{numericvalue to scale the traffic volumen, defauld is 100}

\item{mc.cores}{how much cores should be used for parallelisation, default is
one core less you maximum number of detected cores.}
}
\value{
a regular grid which covered the plz areas
}
\description{
make_grid
}
\examples{
make_grid(plz = 80993, lambda = 0.01)
}
