% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_dataset.R
\docType{data}
\name{mini_dataset}
\alias{mini_dataset}
\title{mini_dataset}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 420 rows and 17 columns.}
\usage{
data(mini_dataset)
}
\description{
A dataset which contains 5 days of PM10 in Munich.
That dataset is only used for purposes.
}
\examples{
data(mini_dataset)
}
\keyword{datasets}
