% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muc_airPol_p1.R
\docType{data}
\name{muc_airPol_p1}
\alias{muc_airPol_p1}
\title{muc_airPol_p1}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 2688 rows and 17 columns.}
\usage{
data(muc_airPol_p1)
}
\description{
A dataset which contains observations of PM10 and a lot of covariables in
munich December 2017.
}
\examples{
data(muc_airPol_p1)
}
\keyword{datasets}
