% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muc_airPol_p1_grid.R
\docType{data}
\name{muc_airPol_p1_grid}
\alias{muc_airPol_p1_grid}
\title{muc_airPol_p1_grid}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 129828 rows and 12 columns.}
\usage{
data(muc_airPol_p1_grid)
}
\description{
A dataset which contains a grid with covariables in
munich December 2017 for predictions of PM10.
}
\examples{
data(muc_airPol_p1_grid)
}
\keyword{datasets}
