% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muc_airPol_p2.R
\docType{data}
\name{muc_airPol_p2}
\alias{muc_airPol_p2}
\title{muc_airPol_p2}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 2604 rows and 17 columns.}
\usage{
data(muc_airPol_p2)
}
\description{
A dataset which contains observations of PM2.5 and a lot of covariables in
munich December 2017.
}
\examples{
data(muc_airPol_p2)
}
\keyword{datasets}
