% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nuts3.R
\docType{data}
\name{NUTS_3}
\alias{NUTS_3}
\title{NUTS_3 to PLZ Mapping}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 8311 rows and 2 columns.}
\usage{
data(NUTS_3)
}
\description{
Maps each German PLZ to a NUTS_3 code
}
\examples{
data(NUTS_3)
}
\keyword{datasets}
