% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_hist.R
\name{optim_hist_para}
\alias{optim_hist_para}
\title{optim_hist_para}
\usage{
optim_hist_para(model_data, data)
}
\arguments{
\item{model_data}{a object as returned by \link{get_model_frame}}

\item{data}{a object as returned by \link{smooth_space_time_variables}}
}
\value{
the optimized parameter in a \link{data.frame}
}
\description{
optimise all parameters to maximize the correlation for the two historical
variables wind and rain
}
