% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_lambda.R
\name{optim_lambda}
\alias{optim_lambda}
\title{Optimize a lambda parameter for the traffic volume}
\usage{
optim_lambda(sensor_data, sensors, roads, validation_plot = TRUE,
  lambda_range = seq(0.001, 0.02, length.out = 10),
  mc.cores = parallel::detectCores() - 1)
}
\arguments{
\item{sensor_data}{data as returned by \link{get_sensor_measured_values}}

\item{sensors}{data as returned by \link{get_sensors}}

\item{roads}{as returned by \link{get_opentransportmap_data}}

\item{validation_plot}{Plot a validation chart? default is TRUE}

\item{lambda_range}{over which range should lambda be optimized}

\item{mc.cores}{how much cores should be used for parallelization, default is
one core less your maximum number of detected cores.}
}
\value{
the lambda which maximizes the correlation between the measured
PM values and the trafficvolumne data
}
\description{
Optimize a lambda parameter for the traffic volume
}
