% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_st_predictions.R
\name{plot.spAirPol.valdidation_set}
\alias{plot.spAirPol.valdidation_set}
\title{plot.spAirPol.valdidation_set}
\usage{
\method{plot}{spAirPol.valdidation_set}(set)
}
\arguments{
\item{set}{a object as returned by \link{get_test_and_training_set}}
}
\description{
Plot the locations of the training and test sensors
}
\examples{
data("mini_dataset")
training_set <- get_test_and_training_set(mini_dataset)
plot(training_set)
}
