% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dt_to_points.R
\name{points_inside_multipolygon}
\alias{points_inside_multipolygon}
\title{points_inside_multipolygon}
\usage{
points_inside_multipolygon(points, multipolygon)
}
\arguments{
\item{points}{points of type sfc}

\item{multipolygon}{polygon of type MULTIPOLYGON}
}
\value{
points which are inside the multipolygon
}
\description{
check which points are contained by the multipolygon and return them.
}
