% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model.R
\name{predict.stAirPol.model}
\alias{predict.stAirPol.model}
\title{predict.stAirPol.model}
\usage{
\method{predict}{stAirPol.model}(spT.model, newdata, training_set = NULL, ...)
}
\arguments{
\item{spT.model}{a model as returned by \link{fit_sp_model}}

\item{newdata}{a dataset with informations of the covariable}

\item{training_set}{an object generated by \link{get_test_and_training_set},
if those an object is delivered, the newdata argument will reduced to the
in training_set specified test sensors.}

\item{...}{additional arguments}
}
\value{
the newdata with added columns of the prediction calculations
}
\description{
predict.stAirPol.model
}
\examples{
data("mini_dataset")
mini_dataset <- clean_model_data(mini_dataset)
formula = value ~ humi + temp + rainhist + windhist +
  trafficvol + log(sensor_age)
training_set <- get_test_and_training_set(mini_dataset, sampel_size = 0.75,
                                          random.seed = 220292)
model.gp <- fit_sp_model(data = mini_dataset, formula = formula,
                         model = 'GP', training_set = training_set)
pred.gp <- predict(model.gp, mini_dataset, training_set)
}
