% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_split.R
\name{predict_split}
\alias{predict_split}
\title{Title}
\usage{
predict_split(model, new_data, sample_count = 100,
  mc.cores = parallel::detectCores() - 1, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
