% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_st_predictions.R
\name{print.spAirPol.valdidation_set}
\alias{print.spAirPol.valdidation_set}
\title{print.spAirPol.valdidation_set}
\usage{
\method{print}{spAirPol.valdidation_set}(set)
}
\arguments{
\item{set}{a object as returned by \link{get_test_and_training_set}}
}
\value{
Informations about the training set
}
\description{
print.spAirPol.valdidation_set
}
\examples{
data("mini_dataset")
training_set <- get_test_and_training_set(mini_dataset)
print(training_set)
}
