% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_information.R
\name{print.stAirPol.aggregation_information}
\alias{print.stAirPol.aggregation_information}
\title{Print method for class stAirPol.aggregation_information}
\usage{
\method{print}{stAirPol.aggregation_information}(x)
}
\arguments{
\item{x}{object of the class stAirPol.aggregation_information}
}
\description{
Print method for class stAirPol.aggregation_information
}
\examples{
m.agg_infor <- aggregation_information(timeshift = lubridate::hours(2),
aggregation_interval = '8 hours')
print(m.agg_infor)
}
