% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalulate_prediction.R
\name{RMSE}
\alias{RMSE}
\title{RMSE}
\usage{
RMSE(z, zhat)
}
\arguments{
\item{z}{observed values}

\item{zhat}{predicted values}
}
\description{
RMSE
}
\examples{
data("mini_dataset")
mini_dataset <- clean_model_data(mini_dataset)
formula = value ~ humi + temp + rainhist + windhist +
  trafficvol + log(sensor_age)
training_set <- get_test_and_training_set(mini_dataset, sampel_size = 0.75,
                                          random.seed = 220292)
model.gp <- fit_sp_model(data = mini_dataset, formula = formula,
                         model = 'GP', training_set = training_set)
pred.gp <- predict(model.gp, mini_dataset, training_set)
RMSE(pred.gp$value, pred.gp$prediction)
}
