% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dvd_data.R
\name{setup_station_table}
\alias{setup_station_table}
\title{setup_station_table}
\usage{
setup_station_table(path)
}
\arguments{
\item{path}{filepath where the informations should be stored}
}
\description{
setup_station_table
}
