% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_lambda.R
\name{smooth_loaction}
\alias{smooth_loaction}
\title{Smooth the trafficvolumne around one point}
\usage{
smooth_loaction(roads, point, lambda = 0.05)
}
\arguments{
\item{roads}{an object as returned by \link{get_opentransportmap_data}}

\item{point}{on element as returned by \link{format_dt_to_points}}

\item{lambda}{smoothing parameter}
}
\value{
numerical value of the estimated trafficvolumne
}
\description{
Smooth the trafficvolumne around one point
}
