% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_space_time_variables.R
\name{smooth_space_time_variables}
\alias{smooth_space_time_variables}
\title{Spatial Interpolation with Kriging}
\usage{
smooth_space_time_variables(sensor_data, sensors = NULL, grid, agg_info,
  mc.cores = parallel::detectCores() - 1, quantiles_to_use = c(0, 1),
  times_IQR = 2.5, ...)
}
\arguments{
\item{sensor_data}{data as returned by \link{get_sensor_measured_values}}

\item{sensors}{data as returned by \link{get_sensors}}

\item{grid}{an object as returned by \link{make_grid_traffic}}

\item{agg_info}{an object as returnet by \link{aggregation_information}}

\item{mc.cores}{how much cores should be used for parallelization, default is
one core less your maximum number of detected cores.}

\item{times_IQR}{remove outliers, which are more then times_IQR * IQR away
from the median}

\item{...}{arguments passed to \link{automap::autoKrige}}
}
\value{
the interpolated dataset around the delivered grid
}
\description{
The functions takes the delivered informations of sensors and sensor_data
and calculates a spatial interpolation with kriging for each unique
timestamp. For the Kriging we use \link{automap::autoKrige}
}
