% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_db_in_months.R
\name{split_db_in_months}
\alias{split_db_in_months}
\title{Split databases into monthly databases}
\usage{
split_db_in_months(path)
}
\arguments{
\item{path}{filepath where the informations should be stored}
}
\description{
Split databases into monthly databases
}
\examples{
split_db_in_months(path = './')
}
\author{
Volker Schmid
}
