% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensor_values.R
\name{sql_load_sensor_measured_values}
\alias{sql_load_sensor_measured_values}
\title{sql_load_sensor_measured_values}
\usage{
sql_load_sensor_measured_values(sensor_id, date_pattern, variable_type, path)
}
\arguments{
\item{sensor_id}{numeric vector of sensor_ids of the luftdaten project}

\item{date_pattern}{date pattern which specifies the month in the format YYYY_MM}

\item{variable_type}{e.g. P1, P2, temperature, humidity}

\item{path}{filepath where the informations are stored}
}
\value{
all measured values
}
\description{
read all measured values for one date_pattern variable_type combination
}
