% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_to_shape_file.R
\name{aggregate_to_shape_file}
\alias{aggregate_to_shape_file}
\title{Title}
\usage{
aggregate_to_shape_file(prediction, shape, scale_bar_dist = 3)
}
\arguments{
\item{scale_bar_dist}{}
}
\description{
Title
}
\examples{
remove points which does not contain in the shapefile
remove shapes where no gridpoints are detected
Plot the calculated aggregation
}
