% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_datasets.R
\name{combine_datasets}
\alias{combine_datasets}
\title{Combine all calculated datasets}
\usage{
combine_datasets()
}
\value{
nothing, but four objects are assigned to the global environment
}
\description{
Combine all calculated datasets
}
\examples{
combine_datasets()
}
