% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_subintervalls.R
\name{fit_subintervalls}
\alias{fit_subintervalls}
\title{fit_subintervalls}
\usage{
fit_subintervalls(data, grid = NULL, training_set = NULL, unit = "1 week",
  nItr = 5000, nBurn = 1000, return = "prediction", tol = 0.01,
  retry_count = 5, retry_reason_mean = 500, retry_reason_rmse = 2000,
  mc.cores = parallel::detectCores() - 1, ...)
}
\arguments{
\item{data}{Modelling data.frame which contains information about the
covariables and the target variable}

\item{grid}{grid data where a prediction should be calculated or a
training_set}

\item{training_set}{an object generated by \link{get_test_and_training_set},
if those a object is delivered, the data argument will reduced to the
in training_set sepcified training sensors.}

\item{unit}{unit for splitting the dataset, see \link{lubridate::floor_date}
for more informations}

\item{nItr}{Number of MCMC iterations. Default value is 5000.}

\item{nBurn}{Number of burn-in samples. This number of samples will be
discarded before making any inference. Default value is 1000.}

\item{return}{default is 'prediction'}

\item{tol}{Minimum separation distance between any two locations out of
those specified by coords, knots.coords and pred.coords. The default is
0.005. The programme will exit if the minimum distance is less than the
non-zero specified value. This will ensure non-singularity of the
covariance matrices.}

\item{retry_count}{How much retrys should be tryed, if some strange behavior
appears, for example a posteriori Mean above 500 or a RMSE of the prediction
above 2000. Both could be specified seperatly. Default is 5.}

\item{retry_reason_mean}{retry on a posteriori Mean above that value,
default is 500}

\item{retry_reason_rmse}{retry on a RMSE above that value, default is 2000}

\item{mc.cores}{how much cores should be used for parallelisation, default is
one core less you maximum number of detected cores.}

\item{...}{Parameters passed to \link{fit_sp_model}}
}
\value{
the grid with added columns of the prediction calculations, or the
crossvaldiation predictions
}
\description{
Fit the spatio-temporal models on subintervall.
}
\examples{
data("mini_dataset")
mini_dataset <- clean_model_data(mini_dataset)
formula = value ~ humi + temp + rainhist + windhist +
  trafficvol + log(sensor_age)
training_set <- get_test_and_training_set(mini_dataset, sampel_size = 0.75,
                                          random.seed = 220292)
prediction.gp <- fit_subintervalls(data = mini_dataset, formula = formula,
                            model = 'GP', training_set = training_set,
                            unit = '1 day')
retry model fitting and prediction in cause of a strange behavior
sometimes
remove sensors which are always NA
}
