% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_traffic.R
\name{make_data_traffic}
\alias{make_data_traffic}
\title{make_data_traffic}
\usage{
make_data_traffic(sensors, lambda, scale = 100,
  mc.cores = parallel::detectCores() - 1)
}
\arguments{
\item{sensors}{data as returned by \link{get_sensors}}

\item{lambda}{smoothing parameter}

\item{scale}{numeric value to scale the traffic volume, default is 100}

\item{mc.cores}{how much cores should be used for parallelization, default is
one core less your maximum number of detected cores.}
}
\value{
\link{data.table} of the sensors with a new column which contains
informations about the traffic volume
}
\description{
Calculate for each sensor the traffic volume data
}
