% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muc_airPol_p1.R
\docType{data}
\name{muc_airPol_p1}
\alias{muc_airPol_p1}
\title{muc_airPol_p1}
\format{A data table with 2604 rows and 17 variables:

\describe{
  \item{timestamp}{timestamp rounded down}
  \item{sensor_id}{numeric sensor_id}
  \item{lon}{longitudinal value of the position}
  \item{lat}{latitudinal value value of the position}
  \item{variable}{Variable Type, P1 for PM10 or P2 for PM2.5}
  \item{sensor_type}{sensor_type, currently nearly always SDS011}
  \item{location}{location indicator}
  \item{value}{measured value of PM}
  \item{humi}{humiditiy in percent}
  \item{temp}{temperature in degree Celsius}
  \item{rain}{rain in millimeters}
  \item{rainhist}{transformed historical rain}
  \item{wind}{wind velocity in m/s}
  \item{windhist}{transformed historical wind}
  \item{id}{numeric id}
  \item{trafficvol}{calculated trafficvolumne}
  \item{sensor_age}{sensor age in weeks}
}}
\usage{
data(muc_airPol_p1)
}
\description{
A dataset which contains observations of PM10 and a lot of covariables in
Munich December 2017.
}
\examples{
data(muc_airPol_p1)
}
\keyword{datasets}
