% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muc_airPol_p2_grid.R
\docType{data}
\name{muc_airPol_p2_grid}
\alias{muc_airPol_p2_grid}
\title{muc_airPol_p2_grid}
\format{A data table with 129828 rows and 12 variables:

\describe{
  \item{lon}{longitudinal value of the position}
  \item{lat}{latitudinal value value of the position}
  \item{humi}{humiditiy in percent}
  \item{timestamp}{timestamp rounded down}
  \item{temp}{temperature in degree Celsius}
  \item{rain}{rain in millimeters}
  \item{rainhist}{transformed historical rain}
  \item{wind}{wind velocity in m/s}
  \item{windhist}{ transformed historical wind}
  \item{trafficvol}{calculated trafficvolumne}
  \item{sensor_id}{numeric sensor_id}
  \item{sensor_age}{sensor age in weeks}
}}
\usage{
data(muc_airPol_p2_grid)
}
\description{
A dataset which contains a grid with covariables in
Munich December 2017 for predictions of PM2.5.
}
\examples{
data(muc_airPol_p2_grid)
}
\keyword{datasets}
