% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parameters.R
\name{plot.stAirPol.parameters}
\alias{plot.stAirPol.parameters}
\title{plot.stAirPol.parameters}
\usage{
\method{plot}{stAirPol.parameters}(parameter)
}
\arguments{
\item{parmeter}{a object as returned by \link{fit_subintervalls} with the
option \code{return = 'parameters'}}
}
\value{
a ggplot2 object
}
\description{
Plot boxplox for each parameter.
}
