% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plz_shape.R
\docType{data}
\name{plz_shape}
\alias{plz_shape}
\title{PLZ shape file}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 59225 rows and 19 columns.}
\usage{
data(plz_shape)
}
\description{
PLZ shape file
}
\examples{
data(plz_shape)
}
\references{
(\href{http://www.geodatenzentrum.de/ geodaten/})
}
\keyword{datasets}
